[Post-Demo] Migrate Terraform code from project repo  (#30)
Ticket
https://wicmtdp.atlassian.net/browse/PRP-77

Changes
infra/mock_apir path added with the infrastructure for the eligibility screener repo

Context for reviewers
TThe Mock API could benefit from being treated as an atomic, separate standalone service that can talk to multiple services. Right now, the terraform for the Mock API is tracked in a separate repo alongside the eligibility screener. Let’s actually break that apart and package the infrastructure for the Mock API in the Mock API repo.

This ticket should be a migration of the terraform out of the project repo and into the Mock API repo. The end result should be that we should be able to manage AWS via terraform for Mock API.